requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-community
langchain-core
langchain-huggingface
faiss-cpu
pypdf2
sentence-transformers